Add cast of literals to CGFloat to avoid implicit conversion errors

+++ Stripe/STPColorUtils.m
@@ -16,13 +16,13 @@
         // We're using the luma value from YIQ
         // https://en.wikipedia.org/wiki/YIQ#From_RGB_to_YIQ
         // recommended by https://www.w3.org/WAI/ER/WD-AERT/#color-contrast
-        return red * 0.299 + green * 0.587 + blue * 0.114;
+        return red * (CGFloat)0.299 + green * (CGFloat)0.587 + blue * (CGFloat)0.114;
     } else {
         // Couldn't get RGB for this color, device couldn't convert it from whatever
         // colorspace it's in.
         // Make it "bright", since most of the color space is (based on our current
         // formula), but not very bright.
-        return 0.4;
+        return (CGFloat)0.4;
     }
 }
